Minor: Rearranged custom WTForm classes into single module
